// src/types/mock-data.type.ts
export type MockData = {
  titles: string[];
  descriptions: string[];
  genres: string[];
  previews: string[];
  videos: string[];
  actors: string[];
  directors: string[];
  users: string[];
  emails: string[];
  avatars: string[];
  posters: string[];
  bgImages: string[];
  bgColors: string[];
};